fix definition of inverse tree density was wrong
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1278 8aca7d54-2c30-db11-9de9-000461428c89

+++ dings/baum.h
@@ -39,7 +39,7 @@ private:
 	static uint8 forest_count_divisor;
 	static uint8 forest_boundary_blur;
 	static uint8 forest_boundary_thickness;
-	static uint8 forest_inverse_spare_tree_density;
+	static uint16 forest_inverse_spare_tree_density;
 	static uint8 max_no_of_trees_on_square;
 
 	void saee_baum();
